backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
celery[redis]==5.3.4
redis==5.0.1
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
APScheduler==3.10.4
dnspython==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-nmap==0.7.1
ipaddress==1.0.23
validators==0.22.0
reportlab==4.0.7
pandas==2.1.3
python-dateutil==2.8.2
pytz==2023.3
cachetools==5.3.2
tenacity==8.2.3